requirements.txt
django==1.9.2
psycopg2==2.6.1
djangorestframework==3.3.2
django-cors-headers==1.1.0
django-environ==0.4.0
requests==2.9.1
django-import-export==0.4.2
raven==5.10.2
django-localflavor==1.2
django-reversion==1.10.1
django-reversion-compare==0.6.1
diff-match-patch==20121119
Pillow==3.1.1
#dev
django-debug-toolbar==1.4
Werkzeug==0.11.3